libpkgconf: pkg: pkgconf_compare_version: use size_t for strlen result

+++ libpkgconf/pkg.c
@@ -1095,7 +1095,7 @@ pkgconf_compare_version(const char *a, const char *b)
 
 		if (isnum)
 		{
-			int onelen, twolen;
+			size_t onelen, twolen;
 
 			while (*one == '0')
 				one++;
